Centralize edit logic in field type definitions
