+++ packages/dataviews/src/field-types/index.tsx
@@ -40,5 +40,6 @@ export default function getFieldTypeDefinition( type?: FieldType ) {
 
 			return true;
 		},
+		Edit: () => null,
 	};
 }
